Fixed formatting, fixed one of the examples which was failing.

+++ tasks/prepr-grunt.js
@@ -9,21 +9,23 @@ module.exports = function(grunt) {
             files = grunt.file.expand(this.data.src),
             destFolder = this.data.dest,
             logging = this.data.logging || false;
-
+        
         files.forEach(function(file) {
-            if (logging) {
-                console.log("Preprocessing " + file);
+            if (fs.lstatSync(file).isFile()) {
+                if (logging) {
+                    console.log("Preprocessing " + file);
+                }
+
+                var fileContents = fs.readFileSync(file).toString(),
+                    preprocessedFileContents = prepr.preprocess(fileContents, defined),
+                    destFile = destFolder ? destFolder + path.sep + path.basename(file) : file;
+
+                if (logging) {
+                    console.log("Writing to " + destFile);
+                }
+
+                fs.writeFileSync(destFile, preprocessedFileContents);
             }
-
-            var fileContents = fs.readFileSync(file).toString(),
-                preprocessedFileContents = prepr.preprocess(fileContents, defined),
-                destFile = destFolder ? destFolder + path.sep + path.basename(file) : file;
-
-            if (logging) {
-                console.log("Writing to " + destFile);
-            }
-
-            fs.writeFileSync(destFile, preprocessedFileContents);
         });
     });
 };
